+++ src/build_options.rs
@@ -13,6 +13,7 @@ use crate::cargo_toml::CargoTomlMetadata;
 use crate::cargo_toml::CargoTomlMetadataPyo3Pack;
 use crate::BuildContext;
 use crate::CargoToml;
+use crate::Manylinux;
 use crate::Metadata21;
 use crate::PythonInterpreter;
 use crate::Target;
@@ -21,6 +22,21 @@ use crate::Target;
 #[derive(Debug, Serialize, Deserialize, StructOpt, Clone, Eq, PartialEq)]
 #[serde(default)]
 pub struct BuildOptions {
+    /// Whether to use and check for compliance with the manylinux1 tag (1),
+    /// use it but don't check compliance (1-unchecked)
+    /// or use the native linux tag (off)
+    ///
+    /// This option is ignored on all non-linux platforms
+    // The options use those named to support the upcoming manylinux 2010 tag.
+    #[structopt(
+        long = "manylinux",
+        raw(
+            possible_values = r#"&["1", "1-unchecked", "off"]"#,
+            case_insensitive = "true",
+            default_value = r#""1""#
+        )
+    )]
+    pub manylinux: Manylinux,
     #[structopt(short = "i", long = "interpreter")]
     /// The python versions to build wheels for, given as the names of the
     /// interpreters. Uses autodiscovery if not explicitly set.
@@ -41,7 +57,7 @@ pub struct BuildOptions {
     /// directory in the project's target directory
     #[structopt(short = "o", long = "out", parse(from_os_str))]
     pub out: Option<PathBuf>,
-    /// Don't check for manylinux compliance
+    /// [deprecated, use --manylinux instead] Don't check for manylinux compliance
     #[structopt(long = "skip-auditwheel")]
     pub skip_auditwheel: bool,
     /// The --target option for cargo
@@ -62,6 +78,7 @@ pub struct BuildOptions {
 impl Default for BuildOptions {
     fn default() -> Self {
         BuildOptions {
+            manylinux: Manylinux::Manylinux1,
             interpreter: vec![],
             bindings: None,
             manifest_path: PathBuf::from("Cargo.toml"),
@@ -146,6 +163,13 @@ impl BuildOptions {
 
         let rustc_extra_args = split_extra_args(&self.rustc_extra_args)?;
 
+        let manylinux = if self.skip_auditwheel {
+            eprintln!("--skip-auditwheel is deprecated, use --manylinux=1-unchecked");
+            Manylinux::Manylinux1Unchecked
+        } else {
+            self.manylinux
+        };
+
         Ok(BuildContext {
             target,
             bridge,
@@ -156,7 +180,7 @@ impl BuildOptions {
             out: wheel_dir,
             release,
             strip,
-            skip_auditwheel: self.skip_auditwheel,
+            manylinux,
             cargo_extra_args,
             rustc_extra_args,
             interpreter,